feat: change notification sound behavior
The way the notification sounds were playing, they were audible in the
recorded audio. This sometimes lead to the transcription models
erroneously adding a spurious 'o' at the beginning of the transcript.
To prevent this:
- We only start recording after the start notification is done playing.
We shorten the delay before showing the recording UI by the amount of time
playing the audio takes.
- We play the stop notification after recording is stopped (and shorten
the stopping delay so the delay doesn't feel too laggy).