refactor: use file as recorder buffer instead of in-memory
Configuration changes such as rotations cause jetpack composables to lose
all state. We need to somehow persist the audio data in case the screen rotates.
If keeping the audio in memory, the potential overhead of saving and
restoring the actual audio samples is large. If saving the samples to a
file, we only have to save the name of the file. We therefore use a file
to store the samples instead of an in-memory array.