feat: keep recording state over configuration changes like rotations
On Android and Jetpack Compose, a configuration changes destroys and restarts
the current activity. If not properly handled, state can be lost. We have moved the
`microphoneUIState` variable from `microphoneState` to the `AttendiMicrophone` composable
such that it can be saved using `rememberSaveable`.

We use the lifecycle events `ON_PAUSE` and `ON_RESUME` to detect either backgrounding or
configuration changes. On `ON_PAUSE`, recording is stopped. On `ON_RESUME`, recording continues.
Since we have changed the implementation to use a file as a buffer, the audio samples are persisted
over the configuration changes.