feat: allow handling lifecycle events in client
We move the `start` and `stop` functions to `microphoneState` such that
they can be called by plugins to programmatically control starting and stopping
instead of only exposing it through a click.
We add a plugin API `onLifeCycle` that registers a callback that is called
when a lifecycle event such as `ON_PAUSE` or `ON_RESUME` happens. Here,
the caller can decide what should happen at certain key moments of the
activity lifecycle. For example, to stop the recording when the activity is paused.
Currently, the events `ON_START`, `ON_RESUME`, `ON_CREATE` can not be used
on screen rotations, since the `microphoneState` is lost after screen rotation.
This is something we have to try to correct in the furure.