fix: play audio after rotations and more careful audio plugin registration
Before, the audio notification plugin would not play audio after rotations.
This happened since the sounds would be loaded on the first click. However, since
we currently keep track of the first click with `rememberSaveable`, the
callback that loads the sounds is not executed on  configuration changes
such as rotation. Since the the audio notification plugin _is_ re-activated
on rotations but the first click is persisted, the sounds are not loaded the
second time around. We now circumvent this by loading the sounds directly in the
`activate` function.

We also make the plugin a bit more defensive by tracking whether it has
already been activated. There were reports that the start sound would be
played multiple times, so we take extra care to make sure that we don't
register playing sound callbacks multiple times.

fix: wait for start sound in a different way

In some cases, the start sound is played multiple times. This behavior
seems to have started occurring in 0.2.1, when introduced waiting for the
start audio to be finished before continuing. Here, we use a slightly different
approach in the hopes to fix this.

fix: rollback some changes since the problem was not on our side