feat(streaming_example): wait for websocket connection to open
It's possible that the websocket connection takes some time to establish. If we already start
recording, it's possible we send a lot of audio at once, which can result in
undefined behavior. Therefore we want to wait until the websocket connection is established
before we exit this function. To do so, we use a CountDownLatch, which we decrement when the
connection is established.